openvdb/Makefile: Make python module install path a changable variable
The location to the installed python headers and modules are all variables
because they vary from system to system, so it makes sense to have the
install path a variable as well. It can't be the same as the system
location, because some distros need to compile and install to a temp
location before being moved -- Gentoo Linux, for example. Thus, the
temp location can be declared. Destdir cannot be used as it would cause
duplicated directory structure, for example:
DESTDIR=/tmp/usr
PYTHON_LIB_DIR=/usr/lib64/python2.7/site-packages
${DISTDIR}${PYTHON_LIB_DIR}=/tmp/usr/usr/lib64/python2.7/site-packages
DESTDIR needs to have usr in it otherwhise headers would be installed to
/include rather than /usr/include
Also, this makes it possible to compile 'make python' in a loop for producing
Python modules for different versions of Python installed on the system.