FindILMBase, FindOpenEXR: Move Namespace version regex inside IF statement
If namespace versioning is off, then don't try to find the version
strings in IlmBaseConfig.h or OpenEXRConfig.h. On some distros,
those files are dummy files for multi-arch support, which just
contain includes to the real files and there is no version information
in these files. This will cause build failures as the variables
will be blank.

This is compile tested on Gentoo Linux.

Signed-off by: Jonathan Scruggs (j.scruggs@gmail.com)