Suppress to-Python converter registration warnings.
It's not clear to me why to-Python converters for GridType::Ptrs
are registered automatically in some cases and not in others,
but it doesn't seem to be based simply on the Boost version.
This change should at least help to suppress "to-Python converter
already registered" warnings, if not eliminate them completely.

Signed-off-by: Peter Cucka <peter.cucka@dreamworks.com>