Fix VS2019 compile problem.
It was depending on undefined behaviour that happened to work on gcc/clang.
From http://eel.is/c++draft/meta#rqmts-4 we have:
    Unless otherwise specified, the behavior of a program that adds
    specializations for any of the templates specified in this subclause
    [meta] is undefined.
So in Visual C++ 2019, the specializations for std::is_integral and
std::is_arithmetic fails to work at all. To fix, this replace the use of
these type traits with some custom ones instead.

Signed-off-by: jlait <jlait@andorra.sidefx.com>