Add a BuildConfig.h that is auto-confed from the CMake process so we can ship #defines that reflect how OpenVDB was built.
Make USE_EXR such a #define, which in Types.h also gates whether it
uses the internal or external Half.   The internal half is always
compiled so the symbols will be there if someone manages to misconfigure
their .so and their BuildConfig.h; but default will be to have things
just match.

Signed-off-by: jlait <jlait@andorra.sidefx.com>