Add option to disable symbol export
By default non-windows builds of openvdb export all
symbols marked with OPENVDB_API in the code.
This is almost always the correct approach, however when
distributing a library as a binary that uses openvdb in a
context where users of that library might also use different
versions of openvdb (including in development versions with
the same ABI) it can cause symbol collisions in the process.

This changes adds a CMake option and compiler define
OPENVDB_NO_SYMBOL_EXPORT, which removes the "default" visibility
for OPENVDB_API symbols.

In effect this imitates the windows static symbol export behavior
for non-windows static builds.

This only makes sense when building openvdb
 - as a static library (a shared library built with this flag
 would be unusable)
 - with symbols hidden (without this the symbols would have
 "default" visibility anyway)

Signed-off-by: Tristan Barback <tristan.barback@autodesk.com>