Integrate Imath and OpenEXR v3.1
All components seem to compile and tests succeed, although there may
be other constraints that aren't accounted for. I haven't built the
Houdini or Maya plugins, so I'm not sure if they introduce other
complexity.

Imath_ROOT/OpenEXR_ROOT and/or CMAKE_PREFIX_PATH point the build to
specific installations, but in lieu of pre-installed dependencies,
this uses CMake's FetchContent to automatically download and build
Imath and OpenEXR from GitHub, so it's not essential for these
dependencies to exist prior to building. This is the strategy that
OpenEXR 3.x uses to build Imath.

This replaces all instances of "IlmBase" with "Imath". Imath has no
"components", it is a single library that includes half.h along with
everything else.

This removes entirely the local copy of Half.h and the USE_IMATH_HALF
option. Ideally, it would be better to embrace this as a required
dependency and be free of the extra complexity.

Even if this isn't the 100% solution, hopefully it serves as a
starting point.

Signed-off-by: Cary Phillips <cary@ilm.com>