Prototype of support OpenEXR/Imath v3.x alongside v2.x
This is not almost certainly complete and may still be missing
critical features, but hopefully it serves as another reference in the
steps towards full support for OpenEXR/Imath 3.x.

This will look for Imath first (via -DImath_ROOT= if provided), and if
that fails, it will look for IlmBase (via -DIlmBase_ROOT= if
provided), then set the dependencies appropriately. If the
find_package(OpenEXR) finds v3.x it sets the dependencies
appropriately. Both these use $<TARGET_NAME_IF_EXIST:XXX> that
simultaneously pull both IlmBase and Imath libraries, assuming that
only one set of components will exist, with the other being expanded
to nothing. The same thing accounts for the OpenEXR libraries, which
change from "IlmImf" to "OpenEXR". And Iex and IlmThread have moved
from IlmBase to OpenEXR.

The USE_IMATH_HALF option and the local copy of half.h are unchanged
and should work just as before.

As a separate step, the copy of Half.h should be updated, that's
untouched here.

Signed-off-by: Cary Phillips <cary@ilm.com>