Fix: Memoryalignment in nanovdb_GridData
The fields mGridIndex and mGridCount are missing in the struct, which leads to an 8 bit memory missalignment with the C++ data in the buffer of gridhandler->data().
This caused invalid data in mMap, as this member comes after the missalignmented happened. Artefacts of this bug could be seen in the OpenCL version of the viewer.