Synchronization of these files with H19. This includes CE grid support (From 18.5 but last sync Nano wasn't official yet)
The one real change (which isn't just FYI/reference) is to GEO_VDBTranslator
to ignore empty vdbs when building bounding boxes.  We were enlarging
the final stat by the min/max of the bbox, but for invalid boxes that
obviously will be wrong.  This showed up as the File SOP load-as-info
giving huge bounding boxes if an empty vdb was present.

Signed-off-by: jlait <jlait@andorra.sidefx.com>