Fix windows compilation issues
Current problem: When using the most recent version of oneTBB on windows with MSVC(2019), linking results in an LNK1104 "cannot open file tbb12.lib" error.
Cause: the target e.g. openvdb_shared references tbb.lib. This on the other hand is just a reference to tbb12.lib, but the directory containing that file is not included in additional library directories. Therefore the linker can't find it.
This pull request includes the directory into the additional link directories and therefore lets the linker find the missing library files.

Signed-off-by: Michael Lackner <mlackner@student.tugraz.at>