Replace boost::uid with a simple string.
Unlike the comments, however, we were NOT storing UUID as an ascii
string, we were instead storing a fixed byte array of the ascii
version of the UUID string.  This means we have to continue to save a
uuid-like string to ensure old readers will not fail (the boost::uuid
will just fail if it doesn't like where hyphens are)

Do not attempt to format the UUID as a random UUID, instead just
randomize the entire string; this is not meant to interoperate with
any other UUID system.

Generate the random string purely from std::random_device so we get
maximum entropy.  This can throw system specific exceptions, so catch
those and produce an always-fail state of an empty string.  These
files may not load in older versions as the nil string I think will be
rejected by the boost uuid reader; but I also don't think this can
occur in modern systems with /dev/urandom.

Signed-off-by: jlait <jlait@andorra.sidefx.com>