The openvdb::Mat3::transpose() method returns a transpose of itself, it doesn't do an in-place transpose.
Thus the line

    mat.transpose(); // the 3x3 in nanovdb is transposed relative to openvdb's 3x3

is a no-op, and not actually transposing the matrix.  Since we DO
transpose when we do an OpenToNano; it is important we transpose here
or we can't round-trip.

Unfortunately, it has been this way since the initial commit, so it is
quite likely tooling has already worked around this....

Signed-off-by: Jeff Lait <jlait@andorra.sidefx.com>