Add a CMakeLists for standalone caching of fvdb_test_data (#362)
This adds a CMakeLists.txt that simply uses CPM to download / cache the FVDB test data if it is not already in the CPM_SOURCE_CACHE.

This is run by build.sh when running ctest (build.sh ctest).

Note that the CPM_SOURCE_CACHE needs to be set to the same value as set when building the gtests (or unset on both).


Signed-off-by: Mark Harris <mharris@nvidia.com>

* Add CPM_SOURCE_CACHE to env in fvdb-gtests GHA job

Signed-off-by: Mark Harris <mharris@nvidia.com>

---------

Signed-off-by: Mark Harris <mharris@nvidia.com>