Use Seaube.clangFormat vscode extension for clang-format (#508)
We have been getting intermittent crashes of clangd, and since we have been using the clang-format that is built into the clangd server, when clangd crashes we lose the format-on-save capability of vscode for C++ and CUDA C++ files.

This PR adds settings and recommendation for the vscode extension Seaube.clangFormat which allows specifying a clang-format executable to be used for formatting C++ and CUDA C++ files. This is a fork of the more widely used xaver.clang-format vscode extension because the latter has not released an update in years, in spite of having merged NVIDIA pull requests to support the cuda-cpp language.

After merging, you will need to make sure to install the vscode extension Seaube.clangFormat. VSCode will prompt to install it.

---------

Signed-off-by: Mark Harris <mharris@nvidia.com>